+++ src/body_hash.rs
@@ -22,7 +22,11 @@ use crate::body_extractor_fold::body_extractor_fold;
 ///
 /// # type T = u64;
 /// async fn hash_payload(form: BodyHash<web::Json<T>, Sha256>) -> impl Responder {
-///     web::Bytes::copy_from_slice(form.hash())
+///     if !form.verify_slice(b"correct-signature") {
+///         // return unauthorized error
+///     }
+///
+///     "Ok"
 /// }
 /// ```
 #[derive(Debug, Clone)]
@@ -43,7 +47,13 @@ impl<T, D: Digest> BodyHash<T, D> {
         self.hash.len()
     }
 
-    /// Returns tuple containing body type and owned hash.
+    /// Verifies HMAC hash against provides `tag` using constant-time equality.
+    pub fn verify_slice(&self, tag: &[u8]) -> bool {
+        use subtle::ConstantTimeEq as _;
+        self.hash.ct_eq(tag).into()
+    }
+
+    /// Returns tuple containing body type, raw body bytes, and owned hash.
     pub fn into_parts(self) -> (T, Bytes, Vec<u8>) {
         let hash = self.hash().to_vec();
         (self.body, self.bytes, hash)